LojistikProject/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
mysql-connector-python
